move the prometheus metrics into their own package as child of server
